% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecflow_client.R
\name{ecflow_client}
\alias{ecflow_client}
\alias{ecflow_init}
\alias{ecflow_abort}
\alias{ecflow_complete}
\alias{ecflow_meter}
\alias{ecflow_event}
\alias{ecflow_label}
\alias{ecflow_change}
\title{ecflow_client wrappers}
\usage{
ecflow_init(pid = Sys.getpid())

ecflow_abort(pid = Sys.getpid())

ecflow_complete()

ecflow_meter(name, value)

ecflow_event(n)

ecflow_label(name, value)

ecflow_change(name, value)
}
\arguments{
\item{pid}{process identifier to the current task}

\item{name}{name of the item to manipulate}

\item{value}{desired value for \code{name} to take}
}
\description{
Wrapper functions makeing system calls to the ecflow_client to interact with the ecflow_server
}
\details{
These are very basic wrappers..
}
