% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecflow_suite.R
\name{ecflow_suite}
\alias{ecflow_suite}
\alias{suite}
\alias{family}
\alias{task}
\alias{edit}
\alias{cron}
\alias{label}
\title{ecflow_suite definition}
\usage{
suite(name, ...)

family(name, ...)

task(name, ...)

edit(par, val)

cron(start, end, step)

label(name, val = "''")
}
\arguments{
\item{name}{name of the suite, task, family or label}

\item{...}{nested suite objects}

\item{par}{parameter to set}

\item{val}{value to take}

\item{end}{end of cron period}

\item{step}{step of cron}

\item{value}{desired value for \code{name} to take}

\item{strt}{start of cron period}
}
\description{
Functions for building ecFlow suites
}
\details{
cron times should be provided as strings. Parameter values are escaped if they are character strings.
}
\examples{
## defining a simple suite
tmp <- suite("ecFlowR_test",
             edit("ECF_MICRO","\%"),
             edit("ECF_JOB_CMD","Rscript \%ECF_JOB\% > \%ECF_JOBOUT\% 2>&1"),
             edit("ECF_KILL_CMD","kill -15 \%ECF_RID\%"),
             edit("ECF_CHECK_CMD","ps --pid \%ECF_RID\% -f"),
             edit("ECF_NODE","\%ECF_HOST\%"),
             edit("ECF_STATUS_CMD","ps --pid \%ECF_RID\% -f > \%ECF_JOB\%.stat 2>&1"),
             edit("etste",2),
             task("shouldRun",
                  cron("00:00","23:59","00:01"),
                  label("lastRun"),
                  label("lastSucess")),
             task("shouldFail",
                  cron("00:00","23:59","00:01"),
                  label("lastRun"),
                  label("lastSucess")),
             task("catchFail",
                  cron("00:00","23:59","00:01"),
                  label("lastRun"),
                  label("lastSucess"))
             )
writeLines(tmp)

}
